+++ OpenSim/Region/Framework/Scenes/Scene.Inventory.cs
@@ -1588,7 +1588,7 @@ namespace OpenSim.Region.Framework.Scenes
                 {
                     client.SendInventoryFolderDetails(
                         fold.Owner, folder.ID, fold.RequestListOfItems(),
-                        fold.RequestListOfFolders(), (int)fold.Version, fetchFolders, fetchItems);
+                        fold.RequestListOfFolders(), fold.Version, fetchFolders, fetchItems);
                     return;
                 }
             }
@@ -1632,7 +1632,7 @@ namespace OpenSim.Region.Framework.Scenes
 
                 client.SendInventoryFolderDetails(
                     client.AgentId, folder.ID, contents.Items, contents.Folders,
-                    (int)containingFolder.Version, fetchFolders, fetchItems);
+                    containingFolder.Version, fetchFolders, fetchItems);
             }
         }
