+++ OpenSim/Region/Framework/Scenes/Scene.PacketHandlers.cs
@@ -538,7 +538,7 @@ namespace OpenSim.Region.Framework.Scenes
                 {
                     remoteClient.SendInventoryFolderDetails(
                         fold.Owner, folderID, fold.RequestListOfItems(),
-                        fold.RequestListOfFolders(), (int)fold.Version, fetchFolders, fetchItems);
+                        fold.RequestListOfFolders(), fold.Version, fetchFolders, fetchItems);
                     return;
                 }
             }
